Supplied for Assignment
A set of HTML, JS, CSS file has been created for the stopwatch already. The assignment was to create a lap time button and display the results next to the stopwatch section. Whenever we click the lap button, the lap time has to be displayed and appended to the existing one.